# frontend/kucukdevapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "kucukdevapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/{uid}/semesters/{sid}/lessons": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "List Lessons",
        "description": "List all lessons of a semester with given userID, semesterID",
        "operationId": "listLessonsOfSemester",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List all lessons of a semester",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Lessons Users  Uid  Semesters  Sid  Lessons Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonAPIModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "lessons"
        ],
        "summary": "Create Lesson",
        "description": "Create a lessons for a semester with given userID, semesterID",
        "operationId": "createLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add new lesson into a semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/{uid}/semesters/{sid}/lessons/{lid}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Show Lesson",
        "description": "Get a single lesson with given userID, semesterID and lessonID",
        "operationId": "getSingleLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            },
            "name": "lid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single lesson of a semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonAPIModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "lessons"
        ],
        "summary": "Update Lesson",
        "description": "Update a lesson with given userID, semesterID and lessonID",
        "operationId": "updateLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            },
            "name": "lid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update a lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "lessons"
        ],
        "summary": "Delete Lesson",
        "description": "Delete a lesson with given userID, semesterID and lessonID",
        "operationId": "deleteLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            },
            "name": "lid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/{uid}/semesters/{sid}/lessons/{lid}/absences": {
      "post": {
        "tags": [
          "lessons"
        ],
        "summary": "Create Absence",
        "description": "Create an absence for a lesson with given userID, semesterID and lessonID",
        "operationId": "createAbsence",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            },
            "name": "lid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonAbsenceModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add absence into a lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "lessons"
        ],
        "summary": "Delete Absence",
        "description": "Delete an absence from a lesson with given userID, semesterID and lessonID",
        "operationId": "deleteAbsence",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            },
            "name": "uid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            },
            "name": "sid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            },
            "name": "lid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonAbsenceModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Delete absence into a lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "LessonAPIModel": {
        "title": "LessonAPIModel",
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "instructor": {
            "title": "Instructor",
            "type": "string"
          },
          "absenceLimit": {
            "title": "Absencelimit",
            "type": "integer"
          },
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          },
          "absences": {
            "title": "Absences",
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "name": "Algebra",
          "instructor": "Jack Joe",
          "absenceLimit": 21,
          "slots": [
            [
              2,
              7,
              0
            ],
            [
              2,
              8,
              0
            ]
          ],
          "absences": [
            [
              0,
              2,
              7,
              0
            ],
            [
              0,
              2,
              8,
              0
            ]
          ]
        }
      },
      "LessonAbsenceModel": {
        "title": "LessonAbsenceModel",
        "required": [
          "absence"
        ],
        "type": "object",
        "properties": {
          "absence": {
            "$ref": "#/components/schemas/AbsenceModel"
          }
        },
        "example": {
          "absence": {
            "week": 0,
            "day": 2,
            "hour": 7,
            "isLab": 0
          }
        }
      },
      "LessonModel": {
        "title": "LessonModel",
        "required": [
          "name",
          "instructor",
          "absenceLimit",
          "slots"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "minLength": 1,
            "type": "string"
          },
          "instructor": {
            "title": "Instructor",
            "minLength": 1,
            "type": "string"
          },
          "absenceLimit": {
            "title": "Absencelimit",
            "minimum": 0,
            "type": "integer"
          },
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotModel"
            }
          },
          "absences": {
            "title": "Absences",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbsenceModel"
            },
            "default": []
          }
        },
        "example": {
          "name": "Algebra",
          "instructor": "Jack Joe",
          "absenceLimit": 21,
          "slots": [
            {
              "day": 2,
              "hour": 7,
              "isLab": 0
            },
            {
              "day": 2,
              "hour": 8,
              "isLab": 0
            }
          ],
          "absences": [
            {
              "week": 0,
              "day": 2,
              "hour": 7,
              "isLab": 0
            },
            {
              "week": 0,
              "day": 2,
              "hour": 8,
              "isLab": 0
            }
          ]
        }
      },
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "MessageCreate": {
        "title": "MessageCreate",
        "required": [
          "_id",
          "message"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "message": "string"
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}